# Conflicts:
#	.idea/deploymentTargetDropDown.xml
#	.idea/misc.xml
#	app/build.gradle

+++ app/src/main/java/com/example/proyecto_tfg/GameItem.kt
@@ -7,5 +7,5 @@ data class GameItem (
         val title: String,
         val status: String,
         val platform: String,
-        val score: Double
+        val score: Int
     )
